requirements.txt
Django==4.2.7
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0

PyDrive==1.3.1

flake8==6.1.0
isort==5.12.0






